refactor: extract common logic.
